Move the existing .svg graphics from Icons to new resources/icons/ folder.
These icons will no longer be embedded in the paclet directly. Instead the WL
Graphics object which represents them will be precomputed by a script, and only
those expressions will be embedded in the paclet.

This is a workaround for the fact that ResourceFunction["SVGImport"] is somewhat
slow, and causes a noticable lag the first time this resources are parsed in a
session.