Add Organizer palette to the system Palettes menu
Add FrontEnd/Palettes/Organizer.nb and declare a "FrontEnd" extension.
Now when the Organizer paclet is installed in the system, an "Organizer"
palette will appear in the system Palettes menu list.

Overhaul how the palette is loaded/refreshed. The Notebook[{..}] source for
the palette is a very simple:

	Notebook[
		{Cell["Uninitialized Organizer Palette", "Text"]},
		TaggingRules -> {"CG:Organizer" -> {"DocumentType" -> "MainOrganizerPalette"}},
		Initialization :> (
			Needs["Organizer`"];
			Organizer`Utils`HandleUIFailure[
				Organizer`RefreshOrganizerPalette[EvaluationNotebook[]]
			]
		),
		... (* Additional display options *)
	]

The use of Initialization causes the content of the Organizer palette to be
computed fresh every time it is opened.

Entirely remove the old mechanism based on storing a PersistentValue[.., "FrontEndSession"]
for the Organizer palette NotebookObject.