Workaround issue with using dialog inputs to initialize the Organizer
Dialog inputs (SystemDialogInput[..] and ChoiceDialog) cannot be used
during a pre-emptive evaluation. This means that it is not possible to
interactively prompt the user to setup the Organizer during
the `Initialization :> ..` logic for the Organizer palette.

The workaround is to change the Initialization logic for the palette
to show the user a "Setup Organizer" button with `Method -> "Queued".
When the user clicks that button, a non-pre-emptive evaluation is performed,
and we can use dialog inputs to query the user in the normal way.

* Use $SynchronousEvaluation instead of $DynamicEvaluation