feature: Add Organizer.nb stylesheet + new `/` and `[` style key mappings (#47)
* cleanup: Factor out common docked cell logic to new MakeOrganizerDockedCells[..] helper function

* feature: Add new 'ConnorGray/Organizer.nb' stylesheet

Previously, Organizer styles were "burned in" as inline styles in
any notebook created by Organizer. This has the significant disadvantage
of making Organizer notebooks difficult to update, because changes to
the Organizer styles won't automatically be reflect in notebooks created
by previous version of Organizer.

Using a stylesheet means that changes to the stylesheet will automatically be
reflected in any Organizer notebook. (Though, some care is required to ensure
that no backwards-incompatible changes are made to the Organizer.nb stylesheet.)

* feature: Add '[' and '/' style key mappings

* Add dependency on ConnorGray/CellInsertionMenu to support the '/' style key mapping
  - This paclet will be installed automatically when ConnorGray`Organizer` is loaded.

* '[' is a convenience for inserting a new TODO style cell