polishs: Inline the Workspace... switching buttons to remove extra button click (#58)
Switching workspaces is relatively common, but the typical Organizer user is
going to have a very small number of workspaces. Putting the Workspace switcher
buttons in a submenu menus it is a minimum of 3 clicks to switch workspaces.

By inlining these buttons into the parent command menu, a button click is reduced.

Also fixed spacing issue in the command button menu.